BUGFIX: Use ContentDimension preset name in EelHelper::findRandomRecordForGroup
In ReceiverGroupGenerationService::generate the preset name is used in the
filename but the EEL helper use the first value of the dimension, so by ex.
if the preset name is 'en' but the first value 'en_EN', file are generated
with the suffix 'en' but the EEL helper try to load file with 'en_EN' and
found nothing.